Extract shared retry metrics test support into a dedicated module
Move `CounterEvent`, `RecordingMetrics`, and `RecordedCounter` from the
unit test and integration test into `src/github/test_support`, gated with
`#[cfg(any(test, feature = "internal"))]`, so both suites use the same
recorder implementation without duplication.
